services/old_vlm_server/pyproject.toml
[project]
name = "vlm_server"
version = "0.1.0"
description = "VLM Inference Service"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.76.0",
    "huggingface-hub>=0.36.0",
    "lib",
    "modal>=1.2.4",
    "python-dotenv>=1.0.0",
    "tokenizers>=0.22.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.1",
    "torch>=2.9.1",
    "google-genai>=1.55.0",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "accelerate>=1.12.0",
    "torchvision>=0.24.1",
    "tokenizers>=0.22.1",
]

[project.scripts]
vlm-server = "vlm_server.server:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vlm_server"]